ai-analysis-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
PyPDF2==3.0.1
aiofiles==23.2.0
asyncio-throttle==1.0.2
pymysql>=1.1.0
requests>=2.31.0
pycryptodome>=3.23.0
pymysql>=1.1.0